senti-backend/requirements.txt
fastapi
uvicorn
websockets
google.generativeai
openai
aiohttp
python-dotenv